refactor(notification): group generators
