+++ src/main/java/sleppynavigators/studyupbackend/domain/notification/generator/ChallengeCompleteNotificationMessageGenerator.java
@@ -1,13 +1,14 @@
-package sleppynavigators.studyupbackend.domain.notification;
+package sleppynavigators.studyupbackend.domain.notification.generator;
 
 import org.springframework.stereotype.Component;
 import sleppynavigators.studyupbackend.domain.event.ChallengeCompleteEvent;
 import sleppynavigators.studyupbackend.domain.event.EventType;
+import sleppynavigators.studyupbackend.domain.notification.NotificationMessage;
 
 @Component
 public class ChallengeCompleteNotificationMessageGenerator implements
-    NotificationMessageGenerator<ChallengeCompleteEvent> {
-    
+        NotificationMessageGenerator<ChallengeCompleteEvent> {
+
     private static final String TITLE = "챌린지 완료 알림";
     private static final String MESSAGE_FORMAT = "%s님이 '%s' 챌린지를 완료했습니다. (%.2f%% 달성)";
