+++ src/main/java/sleppynavigators/studyupbackend/domain/notification/generator/NotificationMessageGeneratorFactory.java
@@ -1,24 +1,23 @@
-package sleppynavigators.studyupbackend.domain.notification;
+package sleppynavigators.studyupbackend.domain.notification.generator;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Component;
-import sleppynavigators.studyupbackend.domain.chat.systemmessage.SystemMessageGenerator;
 import sleppynavigators.studyupbackend.domain.event.EventType;
 import sleppynavigators.studyupbackend.domain.event.NotificationEvent;
-import sleppynavigators.studyupbackend.domain.event.SystemMessageEvent;
 
 @Component
 public class NotificationMessageGeneratorFactory {
+
     private final Map<EventType, NotificationMessageGenerator<?>> generatorMap;
 
-    public NotificationMessageGeneratorFactory(List<NotificationMessageGenerator<? extends NotificationEvent>> generators) {
+    public NotificationMessageGeneratorFactory(
+            List<NotificationMessageGenerator<? extends NotificationEvent>> generators) {
         this.generatorMap = generators.stream()
                 .collect(Collectors.toMap(
-                    NotificationMessageGenerator::supportedEventType,
-                    generator -> generator
-                ));
+                        NotificationMessageGenerator::supportedEventType,
+                        generator -> generator));
     }
 
     @SuppressWarnings("unchecked")
